+++ http/src/test/scala/com/twitter/finatra/http/EmbeddedHttpServer.scala
@@ -11,6 +11,7 @@ import com.twitter.finatra.jackson.ScalaObjectMapper
 import com.twitter.inject.conversions.map._
 import com.twitter.inject.server.{EmbeddedHttpClient, EmbeddedTwitterServer, Ports}
 import com.twitter.util.{Duration, Memoize}
+import java.lang.annotation.Annotation
 import scala.collection.JavaConverters._
 
 /**
@@ -1044,6 +1045,30 @@ class EmbeddedHttpServer(
       withJsonBodyNormalizer
     )
   }
+
+  // java-forwarder methods
+  override final def bindClass[T](clazz: Class[T], instance: T): this.type =
+    super.bindClass[T](clazz, instance)
+
+  // java-forwarder methods
+  override final def bindClass[T](clazz: Class[T], annotation: Annotation, instance: T): this.type =
+    super.bindClass[T](clazz, annotation, instance)
+
+  // java-forwarder methods
+  override final def bindClass[T, Ann <: Annotation](clazz: Class[T], annotationClazz: Class[Ann], instance: T): this.type =
+    super.bindClass[T, Ann](clazz, annotationClazz, instance)
+
+  // java-forwarder methods
+  override final def bindClass[T, U <: T](clazz: Class[T], instanceClazz: Class[U]): this.type =
+    super.bindClass[T, U](clazz, instanceClazz)
+
+  // java-forwarder methods
+  override final def bindClass[T, U <: T](clazz: Class[T], annotation: Annotation, instanceClazz: Class[U]): this.type =
+    super.bindClass[T, U](clazz, annotation, instanceClazz)
+
+  // java-forwarder methods
+  override final def bindClass[T, Ann <: Annotation, U <: T](clazz: Class[T], annotationClazz: Class[Ann], instanceClazz: Class[U]): this.type =
+    super.bindClass[T, Ann, U](clazz, annotationClazz, instanceClazz)
 
   /* Private */