+++ inject/inject-app/src/main/scala/com/twitter/inject/app/internal/TwitterTypeConvertersModule.scala
@@ -1,13 +1,26 @@
 package com.twitter.inject.app.internal
 
 import com.twitter.inject.TwitterModule
-import com.twitter.util.{Duration => TwitterDuration}
-import org.joda.time.Duration
+import com.twitter.inject.app.internal.converters.{
+  InetSocketAddressTypeConverter,
+  LocalTimeTypeConverter,
+  StorageUnitTypeConverter,
+  TwitterDurationTypeConverter,
+  TwitterTimeTypeConverter
+}
+import com.twitter.util.{Duration, StorageUnit, Time}
+import java.net.InetSocketAddress
+import java.time.LocalTime
+import org.joda.{time => joda}
 
 private[app] object TwitterTypeConvertersModule extends TwitterModule {
 
   override def configure(): Unit = {
-    addTypeConverter[Duration](JodatimeDurationTypeConverter)
-    addTypeConverter[TwitterDuration](TwitterDurationTypeConverter)
+    addTypeConverter[joda.Duration](JodatimeDurationTypeConverter)
+    addTypeConverter[LocalTime](LocalTimeTypeConverter)
+    addTypeConverter[Time](TwitterTimeTypeConverter)
+    addTypeConverter[Duration](TwitterDurationTypeConverter)
+    addTypeConverter[StorageUnit](StorageUnitTypeConverter)
+    addTypeConverter[InetSocketAddress](InetSocketAddressTypeConverter)
   }
 }