+++ inject/inject-thrift-client/src/main/scala/com/twitter/inject/thrift/ThriftClientException.scala
@@ -2,9 +2,9 @@ package com.twitter.inject.thrift
 
 import com.twitter.inject.thrift.utils.ThriftMethodUtils._
 import com.twitter.inject.utils.ExceptionUtils._
-import com.twitter.scrooge.ThriftMethod
+import com.twitter.scrooge.ThriftMethodIface
 
-case class ThriftClientException(clientLabel: String, method: ThriftMethod, cause: Throwable)
+case class ThriftClientException(clientLabel: String, method: ThriftMethodIface, cause: Throwable)
     extends Exception(cause) {
 
   override def toString = {