+++ inject/inject-thrift/src/main/scala/com/twitter/inject/thrift/utils/ThriftMethodUtils.scala
@@ -1,11 +1,11 @@
 package com.twitter.inject.thrift.utils
 
-import com.twitter.scrooge.ThriftMethod
+import com.twitter.scrooge.ThriftMethodIface
 import com.twitter.util.Memoize
 
 object ThriftMethodUtils {
 
-  val prettyStr = Memoize { method: ThriftMethod =>
+  val prettyStr = Memoize { method: ThriftMethodIface =>
     method.serviceName + "." + method.name
   }
 }